iTodo Completed

+++ Todo/Views/ListView.swift
@@ -10,18 +10,26 @@ import SwiftUI
 struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     var body: some View {
-        List {
-            ForEach(listViewModel.items, id: \.id) { newitem in
-                ListRowView(item: newitem).contentShape(Rectangle())
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItemStatus(item: newitem)
-                        }
-                    }
-            }.onDelete(perform: listViewModel.DeleteItem)
-                .onMove(perform: listViewModel.MoveItem)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                EmptyItemsView().transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items, id: \.id) { newitem in
+                        ListRowView(item: newitem).contentShape(Rectangle())
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItemStatus(
+                                        item: newitem
+                                    )
+                                }
+                            }
+                    }.onDelete(perform: listViewModel.DeleteItem)
+                        .onMove(perform: listViewModel.MoveItem)
+                }
+                .listStyle(PlainListStyle())
+            }
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
             leading: EditButton(),
